requirements.txt
telethon==1.28.5
tqdm==4.66.1
python-dotenv==1.0.0
Pillow==10.0.0  # For image handling, provides a replacement for imghdr
